requirements.txt
requests~=2.25.1
jsonpickle~=1.4.2
PyYAML~=5.3.1
filetype~=1.0.7
polib~=1.1.1